% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch1.R
\name{batch1}
\alias{batch1}
\title{Run batch analysis for single trait in breedR}
\usage{
batch1(data, FMod, RMod = NULL, geneticM = NULL, SpM = NULL,
  pformula = NULL)
}
\arguments{
\item{data}{aim dataset}

\item{FMod}{Fixed mode,should be 'y~1+fixed.factors'.}

\item{RMod}{Randomed mode, should be '~random.factors'.}

\item{geneticM}{Randomed terms for tree model, details see example.}

\item{SpM}{Spatial error terms. details see example.}

\item{pformula}{formula for h2 (or corr).}
}
\value{
the result is returned directly.
}
\description{
\code{batch1} This function carries out batch analysis for  
mult-trait with same model and also output heritability etc. in breedR package.
}
\details{
breedR batch analysis
}
\examples{
library(breedR)
library(breedRPlus)
library(tidyr)

data(douglas)
S3<-subset(douglas,site=='s3')
#summary(S3);str(S3)

S3a<-filter(S3,is.na(dad)) # hs
S3a<-transform(S3a,Mum=factor(mum))
S3a<-droplevels(S3a)
names(S3a)[7:8]<-c('x1','y1')

df<-tidyr::gather(S3a,key=Trait,y,c(-1:-8,-12,-14:-16))
#str(df)


# for parent model
fixed = y ~ 1+orig
random1=~Mum+block
pformula1=h2~4*V1/(V1+V3)

mm<-plyr::ddply(df,'Trait',
    function(dat) batch1(dat,FMod=fixed,
                             RMod=random1,
                             pformula=pformula1)
                             )
#result                             
mm                      

# for tree model
random2=~block
genetic=list(model = 'add_animal',
             pedigree = S3a[,1:3],
             id = 'self')
pformula2=h2~V2/(V2+V3)

mm1<-plyr::ddply(df,'Trait',
    function(dat) batch1(dat,FMod=fixed,
                             RMod=random2,
                             geneticM=genetic,
                             pformula=pformula2)
                             )

#result                             
mm1

  
}
\references{
breedRPlus website:https://github.com/yzhlinscau/
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
