% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breedRPlus.R
\docType{package}
\name{breedRPlus-package}
\alias{breedRPlus-package}
\alias{breedRPlus}
\title{Augments the Use of 'breedR' in Fitting Mixed Models.}
\description{
Augments the Use of 'breedR' in Fitting Mixed Models.
}
