% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.data_prep.R
\name{load_plink}
\alias{load_plink}
\title{Load PLINK files into a bigSNP object and optionally align phenotype}
\usage{
load_plink(
  bed_file,
  bim_file,
  fam_file,
  pheno = NULL,
  id_col_name = NULL,
  impute_method = "mode",
  backingfile = tempfile()
)
}
\arguments{
\item{bed_file}{Path to the .bed file.}

\item{bim_file}{Path to the .bim file.}

\item{fam_file}{Path to the .fam file.}

\item{pheno}{Optional data.frame of phenotypes (must contain id_col_name). Default NULL.}

\item{id_col_name}{Name of the column in pheno containing individual IDs.}

\item{impute_method}{The imputation method. Default is "mode".}

\item{backingfile}{Path for the 'bigSNP' backing file.}
}
\value{
A list containing:
\item{snp_obj}{The bigSNP object.}
\item{pheno}{Phenotype data aligned to genotypes (if provided).}
}
\description{
Load PLINK files into a bigSNP object and optionally align phenotype
}
