% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{Count error for h2 in breedR}
\usage{
\method{pin}{remlf90}(object, formula, signif = FALSE, digit = 5,
  vres = FALSE)
}
\arguments{
\item{object}{an object of breedR result}

\item{formula}{formula for h2 (or corr, not now)}

\item{signif}{Index to output signif levels, F(default) for non-signif.}

\item{digit}{a number to control decimal point,5(default).}

\item{vres}{Index(T) to return results in vectors, F(default) for direct results.}
}
\value{
the result is returned directly.
}
\description{
\code{pin} This function counts standard error(se) for heritability(h2) and corr value 
and also outputs significent level for corr value in breedR package.
}
\details{
counts standard error(se) for heritability(h2)
}
\examples{
library(breedR)
library(breedRPlus)

data(globulus)
res.animal <- remlf90(fixed = phe_X ~ 1,
                      random = ~ gg,
                      genetic = list(model = 'add_animal',
                      pedigree = globulus[, 1:3],
                      id = 'self'),
                      data = globulus)
                      
pin(res.animal, h2~V2/(V1+V2+V3))

}
\references{
breedRPlus website:https://github.com/yzhlinscau/
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
