% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1.R
\name{plot1}
\alias{plot1}
\title{Test trait's norm for breedR object}
\usage{
\method{plot1}{remlf90}(object, mulT = FALSE)
}
\arguments{
\item{object}{an object of BreedR result}

\item{mulT}{multi-trait model(default, FALSE).}
}
\value{
the result is returned directly.
}
\description{
\code{plot1} This function Test trait's norm for breedR object.
}
\details{
Test trait's norm for BreedR object,similar to asreml.
}
\examples{
library(breedR)
library(breedRPlus)

res.animal <- remlf90(fixed = phe_X ~ 1,
                      random = ~ gg,
                      genetic = list(model = 'add_animal',
                      pedigree = globulus[, 1:3],
                      id = 'self'),
                      data = globulus)
plot1(res.animal)

}
\references{
breedRPlus website:https://github.com/yzhlinscau/
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
