% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.bayes_utils.R
\name{run_bglr}
\alias{run_bglr}
\title{Run Bayesian Genomic Prediction using BGLR}
\usage{
run_bglr(
  obj,
  pheno,
  trait_name,
  id_col_name,
  model_type = "BayesA",
  n_iter = 5000,
  burn_in = 1000
)
}
\arguments{
\item{obj}{A bigSNP object containing genotype data (must have $genotypes and $fam).}

\item{pheno}{Data frame containing phenotype data with individual IDs.}

\item{trait_name}{Character string naming the trait column in pheno.}

\item{id_col_name}{Character string naming the ID column in pheno.}

\item{model_type}{Bayesian model type. Options: "BayesA", "BayesB", "BayesC", "BL" (Bayesian Lasso). Default "BayesA".}

\item{n_iter}{Number of MCMC iterations. Default 5000.}

\item{burn_in}{Number of burn-in iterations. Default 1000.}
}
\value{
A list containing:
\item{gebv}{Data frame with individual IDs and genomic estimated breeding values.}
\item{snp_effects}{Data frame with SNP effects.}
\item{residual_var}{Residual variance estimate.}
\item{model_fit}{The full BGLR model fit object.}
}
\description{
Fits a Bayesian regression model for genomic prediction using the BGLR package.
Supports various Bayesian models including BayesA, BayesB, BayesC, and BL.
}
