% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.gblup_utils.R
\name{run_gblup}
\alias{run_gblup}
\title{Run GBLUP and optionally predict for all individuals (phenotyped + unphenotyped)}
\usage{
run_gblup(
  qc_results,
  trait_name,
  id_col_name,
  fixed_effects = ~1,
  drop_missing = TRUE,
  predict_all = FALSE,
  tol_diag = 1e-06
)
}
\arguments{
\item{qc_results}{List with elements:
snp_obj (bigSNP), pheno (data.frame aligned one-row-per-genotype), optional keep_ind (row indices)}

\item{trait_name}{Trait column name (character)}

\item{id_col_name}{ID column name (character)}

\item{fixed_effects}{Formula for fixed effects (default ~1)}

\item{drop_missing}{If TRUE (default) drops individuals with NA phenotype for fitting.}

\item{predict_all}{If TRUE (default FALSE) compute GEBVs for all individuals (including those with missing phenotypes).}

\item{tol_diag}{small ridge added to diagonal for numerical stability (default 1e-6)}
}
\value{
list with elements: gebv (data.frame ID, gebv for requested animals), variances, model_fit, G (full G over keep_ind)
}
\description{
Run GBLUP and optionally predict for all individuals (phenotyped + unphenotyped)
}
