% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.ml.R
\name{run_ml_model}
\alias{run_ml_model}
\title{Run Machine Learning Model for Genomic Prediction}
\usage{
run_ml_model(pheno, geno, trait_name, id_col_name, model_type = "RF", ...)
}
\arguments{
\item{pheno}{Data frame with phenotype data}

\item{geno}{Numeric genotype matrix (rownames = individual IDs)}

\item{trait_name}{Column name of trait in pheno}

\item{id_col_name}{Column name of sample IDs in pheno}

\item{model_type}{One of "RF", "XGB", "MLP", "CNN", or "PLS"}

\item{...}{Additional arguments passed to train_ml_model}
}
\value{
List containing:
\item{gebv}{Data frame with IDs and predicted GEBVs}
\item{model_fit}{The trained model object}
}
\description{
Trains a machine learning model and returns genomic estimated breeding values (GEBVs).
}
