% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var}
\alias{var}
\title{Output variance component for BreedR object}
\usage{
\method{var}{remlf90}(object, mulT = FALSE)
}
\arguments{
\item{object}{an object of breedR result}

\item{mulT}{multi-trait model(default, FALSE).}
}
\value{
the result is returned directly.
}
\description{
\code{var} This function Output variance component for BreedR object.
}
\details{
Output variance component for breedR object,similar to asreml.
}
\examples{
library(breedR)
library(breedRPlus)

res.animal <- remlf90(fixed = phe_X ~ 1,
                      random = ~ gg,
                      genetic = list(model = 'add_animal',
                      pedigree = globulus[, 1:3],
                      id = 'self'),
                      data = globulus)
var(res.animal)

}
\references{
breedRPlus website:https://github.com/yzhlinscau/
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
